Use Int instead of UInt8 for archetype identifiers (#327)

+++ src/lock.jl
@@ -7,14 +7,14 @@ function _Lock()
     _Lock(_BitPool(), 0)
 end
 
-function _lock(lock::_Lock)::UInt8
+function _lock(lock::_Lock)::Int
     l = _get_bit(lock.pool)
     lock.lock_bits |= UInt64(1) << (l - 1)
     return l
 end
 
-function _unlock(lock::_Lock, b::UInt8)
-    if ((lock.lock_bits >> (b - 1)) & 0x01) == 0
+function _unlock(lock::_Lock, b::Int)
+    if ((lock.lock_bits >> (b - 1)) & UInt64(0x01)) == 0
         throw(
             InvalidStateException(
                 "unbalanced unlock. Did you close a query that was already iterated?",
