+++ src/vec_map.jl
@@ -10,14 +10,14 @@ function _VecMap{T,M}() where {T,M}
     _VecMap{T,M}(Vector{T}(undef, 5), _MutableMask{M}())
 end
 
-function _get_map(m::_VecMap, index::UInt8)
+function _get_map(m::_VecMap, index::Int)
     if !_get_bit(m.used, index)
         return nothing
     end
     @inbounds return m.data[index]
 end
 
-function _set_map!(m::_VecMap{T}, index::UInt8, value::T) where T
+function _set_map!(m::_VecMap{T}, index::Int, value::T) where T
     if length(m.data) < index
         size = (index + _vec_map_chunk_size) & -_vec_map_chunk_size
         resize!(m.data, size)
